requirements.txt
GitPython
nltk
